impl H2G for TE curves
